chore: normalize & fix static rank (#119)

+++ ranking/tests/StaticURLRanker_Sim.cpp
@@ -15,7 +15,7 @@ int main() {
     std::string inFilePath = std::string(config.GetString("static_ranker_in_file").Cstr());
     std::ifstream inFile(inFilePath);
 
-    std::set<std::pair<int, std::string>> urlRankingSet;
+    std::set<std::pair<double, std::string>> urlRankingSet;
 
     for (std::string line; std::getline(inFile, line);) {
         // Empty line
